% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.R
\name{confusion}
\alias{confusion}
\title{Calculate confusion matrix}
\usage{
confusion(truth_name, prob_name, threshold = 0.5)
}
\arguments{
\item{truth_name}{(`character(1L)`) Character containing the name of the vector of 0-1-values
encoded as integer or numeric.}

\item{prob_name}{(`character(1L)`) Character containing the name of the vector of probabilities.}

\item{threshold}{(`numeric(1L)`) Threshold used to transform probabilities into classes (default = 0.5).}
}
\value{
Confusion matrix.
}
\description{
This function calculates the confusion matrix.
}
\author{
Daniel S.
}
