% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.R
\name{dsConfusion}
\alias{dsConfusion}
\title{Calculate the confusion matrix the DataSHIELD servers}
\usage{
dsConfusion(connections, truth_name, pred_name, threshold = 0.5)
}
\arguments{
\item{connections}{(`DSI::connection`) Connection to an OPAL server.}

\item{truth_name}{(`character(1L)`) `R` object containing the models name as character.}

\item{pred_name}{(`character(1L)`) Name of the object predictions should be assigned to.}

\item{threshold}{(`numeric(1L)`) Threshold used to transform probabilities into classes (default = 0.5).}
}
\value{
Confusion matrix for multiple server
}
\description{
Calculate the confusion matrix the DataSHIELD servers
}
\author{
Daniel S.
}
