% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{internDim}
\alias{internDim}
\title{Get data dimension}
\usage{
internDim(symbol = "D")
}
\arguments{
\item{symbol}{(`character(1L)`) \cr
Name of the variable at the DataSHIELD server.}
}
\value{
data dimension (dim).
}
\description{
Get data dimension
}
\author{
Daniel S.
}
