% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{internLength}
\alias{internLength}
\title{Get length of vector}
\usage{
internLength(symbol)
}
\arguments{
\item{symbol}{(`character(1L)`) \cr
Name of the variable at the DataSHIELD server.}
}
\value{
length
}
\description{
Get length of vector
}
\author{
Daniel S.
}
