% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{internMean}
\alias{internMean}
\title{Get mean of vector}
\usage{
internMean(symbol)
}
\arguments{
\item{symbol}{(`character(1L)`) \cr
Name of the variable at the DataSHIELD server.}
}
\value{
mean
}
\description{
Get mean of vector
}
\author{
Daniel S.
}
