% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{internN}
\alias{internN}
\title{Get number of rows}
\usage{
internN(symbol = "D")
}
\arguments{
\item{symbol}{(`character(1L)`) \cr
Name of the variable at the DataSHIELD server.}
}
\value{
Number of rows (nrow).
}
\description{
Get number of rows
}
\author{
Daniel S.
}
