% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{internSum}
\alias{internSum}
\title{Get sum of vector}
\usage{
internSum(symbol)
}
\arguments{
\item{symbol}{(`character(1L)`) \cr
Name of the variable at the DataSHIELD server.}
}
\value{
sum
}
\description{
Get sum of vector
}
\author{
Daniel S.
}
