% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_curve_print.R
\name{print.calibration.curve}
\alias{print.calibration.curve}
\title{Printer for calibration curve objects}
\usage{
\method{print}{calibration.curve}(x, ...)
}
\arguments{
\item{x}{(`list()`) List containing the calibration curve object returned from `dsCalibrationCurve()`.}

\item{...}{Additional parameters (basically none, but CRAN forces us to do this in print).}
}
\description{
Printer for calibration curve objects
}
\author{
Daniel S.
}
